Refactored other components to use new withField
