+++ src/components/Input/Input.types.ts
@@ -4,7 +4,7 @@
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
  */
-import { IFieldComponentProps } from '../Field';
+import { IFieldComponentProps } from '../withField';
 
 /**
  * Props for the Input component
@@ -15,4 +15,6 @@ export interface IInputProps extends IFieldComponentProps {
    */
   // tslint:disable-next-line:no-reserved-keywords
   type: string;
+
+  label: string;
 }
