+++ src/index.ts
@@ -5,7 +5,6 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-export * from './components/Field';
 export * from './components/FieldError';
 export * from './components/FieldGroup';
 export * from './components/FieldLine';
@@ -15,6 +14,7 @@ export * from './components/FormContext';
 export * from './components/FormText';
 export * from './components/Input';
 export * from './components/ValidationSummary';
+export * from './components/withField';
 export * from './components/withForm';
 export * from './components/withValidation';
